+++ CHANGELOG.md
@@ -15,6 +15,7 @@ x.y.z Release notes (yyyy-MM-dd)
 * Use `LocationProvider.currentLocation()` first and `LocationProvider.firstValidLocation()` as fallback to speed up the verify location call in the PWA
 * Refactor `LocationProvider` so that every one-time location e.g. used in `isPoiInRange` or `requestLocalApps` is returned inline
 * Decrease `getConfig` PWA message handling timeout to 5 seconds
+* Set `utm_source` per default and add defined list of default `utm_params` to make sure these params won't be removed
 
 
 7.4.0 Release notes (2021-05-12)